+++ src/boids/behaviors.py
@@ -1,21 +1,14 @@
 """Behaviors here implemented as functions taking a boid and it's neighbors and returning a Vector2 force"""
 
-from .vector2 import Vector2
 import numpy as np
 
-MAX_ALIGNMENT = 10
-
-SEPARATION_STR = 100
-ALIGNEMENT_STR = 1.5
-COHESION_STR = 2
-
-AVOID_BOUNDARY_STR = 400
-AVOID_BOUNDARY_MARGIN = 200
+from .config import config
+from .vector2 import Vector2
 
 
 def avoid_boundary(entities):
     force = np.zeros_like(entities.positions)
-    margin = AVOID_BOUNDARY_MARGIN
+    margin = config.behaviors.boundary.margin
 
     x = entities.positions[:, 0]
     y = entities.positions[:, 1]
@@ -30,7 +23,7 @@ def avoid_boundary(entities):
     force[too_up, 1] = ((margin - y[too_up]) / margin) ** 2
     force[too_down, 1] = -(((y[too_down] - (entities.height - margin)) / margin) ** 2)
 
-    return force * AVOID_BOUNDARY_STR
+    return force * config.behaviors.boundary.strength
 
 
 def sep_ali_coh_numpy(entities, neighbors):
@@ -47,14 +40,14 @@ def sep_ali_coh_numpy(entities, neighbors):
         - entities.positions[has_neighbors]
     )
 
-    # alignement
+    # alignment
     velocities_sums = np.zeros_like(entities.velocities)
     np.add.at(
         velocities_sums, neighbors.sources, entities.velocities[neighbors.targets]
     )
 
-    alignement = np.zeros_like(entities.positions)
-    alignement[has_neighbors] = (
+    alignment = np.zeros_like(entities.positions)
+    alignment[has_neighbors] = (
         velocities_sums[has_neighbors] / neighbors.nb_count[has_neighbors, None]
         - entities.velocities[has_neighbors]
     )
@@ -77,7 +70,7 @@ def sep_ali_coh_numpy(entities, neighbors):
     np.add.at(separation, neighbors.sources, contributions)
 
     return (
-        separation * SEPARATION_STR,
-        cohesion * COHESION_STR,
-        alignement * ALIGNEMENT_STR,
+        separation * config.behaviors.separation.strength,
+        cohesion * config.behaviors.cohesion.strength,
+        alignment * config.behaviors.alignment.strength,
     )